module-4/e-commerce-analytics-data-pipeline/pyproject.toml
[project]
name = "e-commerce-analytics-data-pipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "Joseph Lubandi",email = "lubandi1234@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "faker (>=40.1.0,<41.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "e_commerce_analytics_data_pipeline", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
